pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prd-maker"
version = "0.1.0"
description = "AI-powered Product Requirements Document generator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.28.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-ollama>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "ruff>=0.12.2",
]
